Adjusted CronTabChange

+++ torchk_v1.3a.sh
@@ -70,12 +70,15 @@ lanip=$(uci show | grep lan.ipaddr | cut -d\' -f2)
 # =========
 # Function change crontab on error to every 5 minutes, back to hourly if ok
 CronTabChange() {
-if [ -e /tmp/root.tmp ]; then rm /tmp/root.tmp; fi
-if [ -e /etc/crontabs/root ]; then cat /etc/crontabs/root | grep -v /etc/tor/torchk.sh > /tmp/root.tmp; fi
-if [ "$1" = "ok" ]; then echo "  0 * * * * /etc/tor/torchk.sh" >> /tmp/root.tmp; fi
-if [ "$1" = "error" ]; then echo "*/5 * * * * /etc/tor/torchk.sh" >> /tmp/root.tmp; fi
-cp -f /tmp/root.tmp /etc/crontabs/root
-rm /tmp/root.tmp
+if [ ! "$1" = "" ]; then
+  if [ -e /tmp/root.tmp ]; then rm /tmp/root.tmp; fi
+  if [ -e /etc/crontabs/root ]; then cat /etc/crontabs/root | grep -v /etc/tor/torchk.sh > /tmp/root.tmp;
+    if [ "$1" = "ok" ]; then echo "  0 * * * * /etc/tor/torchk.sh" >> /tmp/root.tmp; fi
+    if [ "$1" = "error" ]; then echo "*/5 * * * * /etc/tor/torchk.sh" >> /tmp/root.tmp; fi
+    cp -f /tmp/root.tmp /etc/crontabs/root
+    rm /tmp/root.tmp
+  fi
+fi
 }
 
 # Function LEDs (on=error or off=OK)
